LPS-179094 Extract ScreenNavigationEntry methods from ObjectDefinitionsFieldsScreenNavigationCategory to ObjectDefinitionsFieldsScreenNavigationEntry, implement getCategoryKey method in ObjectDefinitionsFieldsScreenNavigationEntry and make ObjectDefinitionsFieldsScreenNavigationCategory, extends BaseObjectDefinitionsScreenNavigationCategory
